Fix bugs in the release script. (#100)
Need to set GOPATH properly before building the code.

Copy the release scripts to the Docker image rather than cloning the
repo. This makes it easier to push changes to the releaser because
we can just build the image with local changes. Otherwise we need to
commit the changes first and then build the Docker image.

Need to break out of the loop if only building once

Need to install helm in the releaser image because we use it to build the helm package.

Need to build the docker images using GCB so we don't need to access docker inside the pod running the releaser.

Store the image version in a json file inside the container and print out.

Related to #63