Create separate commands to clone and build the repo (#199)
The goal is to be able to split cloning the repo and building the artifacts into two steps #189

We want cloning the repo to be a separate step because we want to pull release.py from the repo when actually building the code. But in order to invoke release.py from the repo we first have to check it out.

In a follow on PR we will update the E2E test pipeline to use the new commands.