Master trains, workers hang waiting for variable initialization
In this implementation we want to be able to keep some variables
private to the workers, namely those in sections marked
ppo_temporary and simulate_temporary. Sharing variables for the
environments across workers causes problems.

Currently non-chief nodes hang, repeating:
  INFO:tensorflow:Waiting for model to be ready.  Ready_for_local_init_op:
  Variables not initialized: ppo_temporary/episodes/Variable,
  ppo_temporary/episodes/Variable_1, ppo_temporary/episodes/Variable_2,
  ppo_temporary/episodes/Variable_3, ppo_temporary/episodes/Variable_4,
  ppo_temporary/episodes/Variable_5, ppo_temporary/last_action,
  ppo_temporary/last_mean, ppo_temporary/last_logstd, ready: None

Question: How can the chief session, once initialized, query for uninitialized
variables that are private to the worker nodes? If it can't do this,
how can it send a signal to these nodes to initialize their own vars.?
Presumably this should be happening via the ready_for_local_init op.

Presumably because these variables which are private to an individual
worker are not initialized by the chief. It seems we should be initializing
these variables in the local_init_op in response to the,
ready_for_local_init_op, not blocking the local_init_op until these
local variables are initialized.