Rename the `async` keyword argument to `async_req` (#790)
* Rename the `async` keyword argument to `async_req`

The Python Kubernetes client [recently changed](https://github.com/kubernetes-client/python/pull/568) the name of the `async` keyword parameter in its API to `async_req`.

This change makes the corresponding change to our calls to that API, which is necessary for the tf-operator code to work against the latest version of the Kubernetes client code.

* Update kubernetes requirement to >= 7.0.0

This also updates the outdated dependencies in the Pipfile.lock file.