manifests: Move manifests development upstream
As part of the work of wg-manifests for 1.3
(https://github.com/kubeflow/manifests/issues/1735), we are moving manifests
development in upstream repos. This gives the application developers full
ownership of their manifests, tracked in a single place.

This commit copies the manifests for application `TFJob Operator`
from path `apps/tf-training/upstream` of kubeflow/manifests to path
`manifests` of the upstream repo (https://github.com/kubeflow/tf-operator).

Signed-off-by: Yannis Zarkadas <yanniszark@arrikto.com>