Remove `table-logger` dependency
Closes https://github.com/kubeflow/training-operator/issues/1446.

This removes the dependency on `table-logger`. None of the various
alternatives mentioned in that issues were drop-in replacements, so
given the simplicity of the use case here the dependency is replaced
with a small class utilizing `str.format`. Note that the class was
written without referring to the `table-logger` source at all, which
would be a license violation.

I confirmed that you can now run the sample notebook

https://github.com/kubeflow/training-operator/blob/6523d8d2b7a27f1140a05971843cc4da7523694f/sdk/python/examples/kubeflow-tfjob-sdk.ipynb

in modern IPython.