Reconcile the services for the MPIJobs.
The distributed training like torchrun requires setting the
MASTER_ADDR. Adding the service for the workers can help the
workers to recgonize the master (for example worker 0) through a
static DNS name.

This is a compatible change for who want to use naive
torch.distributed method in MPIJob instead of changing the job
definition to PyTorchJob.

Signed-off-by: Peng Gao <peng.gao.dut@gmail.com>